Fix command bug
[bug]Uppercase versions of the commands would have been recognised as valid
commands.

In fix this I chose to break a Java convention and use lower case values
for the commands in the Command enum, this seemed the most elegant and
robust solution.
The convention is a little strange when I think about, as each
enumerated value in an enum is an object, so in a strange way this
change follows the Java convention for object names.